Rename browser translation helper and update docs
