+++ packages/dnb-design-system-portal/src/docs/uilib/components/autocomplete/info.mdx
@@ -120,3 +120,7 @@ You can also set the width directly, but then it has to be defined like so (incl
 ```
 
 <AutocompleteMethods />
+
+## Root Element (React Portal)
+
+The Autocomplete component uses [PortalRoot](/uilib/components/portal-root) internally to render its option list. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
