+++ packages/dnb-design-system-portal/src/docs/uilib/components/date-picker/info.mdx
@@ -47,10 +47,6 @@ Or use UTC constructors:
 const utcDate = new Date(Date.UTC(2025, 0, 1))
 ```
 
-### Root Element (React Portal)
-
-The DatePicker component uses [PortalRoot](/uilib/components/portal-root) internally to render its calendar. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM.
-
 ### Manipulate the days in the calendar view
 
 The callback event `onDaysRender` gives you the possibility to manipulate the "day" object before it gets rendered. This callback will be called many times, both on the first render and on every user interaction, like hover and selection. This means you have to ensure a performant date calculation.
@@ -134,3 +130,7 @@ Additional event return object properties:
   isValidEndDate: boolean, /* Available if `range` is `true` */
 }
 ```
+
+## Root Element (React Portal)
+
+The DatePicker component uses [PortalRoot](/uilib/components/portal-root) internally to render its calendar. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
