+++ packages/dnb-design-system-portal/src/docs/uilib/components/dialog/info.mdx
@@ -54,3 +54,7 @@ To provide custom content to parts of the Dialog, a set of component parts are p
 ### More detailed information
 
 For more details regarding the component functionality, check out the [Modal documentation](/uilib/components/modal).
+
+## Root Element (React Portal)
+
+The Dialog component uses [PortalRoot](/uilib/components/portal-root) internally through Modal to render its content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
