+++ packages/dnb-design-system-portal/src/docs/uilib/components/drawer/info.mdx
@@ -30,6 +30,6 @@ To provide custom content to parts of the Drawer, a set of component parts are p
 
 For more details regarding the component functionality, check out the [Modal documentation](/uilib/components/modal).
 
-### Root Element (React Portal)
+## Root Element (React Portal)
 
-The Drawer component uses [PortalRoot](/uilib/components/portal-root) internally to render its content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM.
+The Drawer component uses [PortalRoot](/uilib/components/portal-root) internally to render its content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
