+++ packages/dnb-design-system-portal/src/docs/uilib/components/modal/info.mdx
@@ -72,10 +72,6 @@ Modal dialogs appear on top of the main content changing the _mode_ of the syste
 
 While it is possible to nest a Modal within another Modal, you as a developer must ensure the browser's back button (alongside the URL path) handles the nested Modal situation. A user should still be able to use the back button as expected and return to the last visited "page".
 
-## Root Element (React Portal)
-
-The Modal component uses [PortalRoot](/uilib/components/portal-root) internally to render its content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM.
-
 ### Z-index
 
 The Modal component is using **3000** as the `z-index`.
@@ -97,3 +93,7 @@ html[data-dnb-modal-active='MODAL-ID'] {
   /* Your css */
 }
 ```
+
+## Root Element (React Portal)
+
+The Modal component uses [PortalRoot](/uilib/components/portal-root) internally to render its content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
