+++ packages/dnb-design-system-portal/src/docs/uilib/components/portal-root/demos.mdx
@@ -10,12 +10,12 @@ import * as Examples from './Examples'
 
 <Examples.PortalRootExample />
 
-### AutoTranslate
+### BrowserTranslate
 
-The `AutoTranslate` helper prevents browser translation tools (e.g. Google Translate) from modifying form component content. It sets `translate="no"` on both the component itself and any portal-rendered content like dropdown lists.
+The `BrowserTranslate` helper prevents browser translation tools (e.g. Google Translate) from modifying form component content. It sets `translate="no"` on both the component itself and any portal-rendered content like dropdown lists.
 
 ```tsx
-import { AutoTranslate } from '@dnb/eufemia/shared'
+import { BrowserTranslate } from '@dnb/eufemia/shared'
 ```
 
-<Examples.AutoTranslateExample />
+<Examples.BrowserTranslateExample />
