+++ packages/dnb-design-system-portal/src/docs/uilib/components/term-definition/info.mdx
@@ -35,4 +35,4 @@ To ensure that the TermDefinition component is accessible, it uses semantic HTML
 
 ## Root Element (React Portal)
 
-The TermDefinition component uses [PortalRoot](/uilib/components/portal-root) internally to render its explanation content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM.
+The TermDefinition component uses [PortalRoot](/uilib/components/portal-root) internally to render its explanation content. See the [PortalRoot documentation](/uilib/components/portal-root) for information on how to control where the portal content appears in the DOM, and for the [BrowserTranslate helper](/uilib/components/portal-root/#browsertranslate-helper) when browser translation tools such as Google Translate should not modify content rendered through PortalRoot.
